Move to SemanticModelExtensions
I'm not sure why this extensions class doesn't already exist but this method is a perfect example of a standalone semantic analysis we can do
Plus I'm doing anything I can to pull isolated things apart into separate classes since the main ones are getting so crowded
